blockchain_contracts: Remove dupe test helper code
